+++ HISTORY.rst
@@ -9,8 +9,17 @@ Unreleased Changes
   memory, a warning is issued or logged. https://github.com/natcap/pygeoprocessing/issues/361
 * Fixed an issue in ``extract_strahler_streams_d8`` where a nodata pixel
   could be mistakenly treated as a stream seed point, ultimately creating
-  a stream feature with no geometry. 
+  a stream feature with no geometry.
   https://github.com/natcap/pygeoprocessing/issues/361
+* Improved ``align_and_resize_raster_stack`` so that rasterization of a vector
+  mask only happens once, regardless of the number of rasters in the stack.
+  In addition, the created mask raster's path may be defined by the caller so
+  that it persists across calls to ``align_and_resize_raster_stack``.
+  https://github.com/natcap/pygeoprocessing/issues/366
+* Improved ``warp_raster`` to allow for a pre-defined mask raster to be
+  provided instead of a vector.  If both are provided, the mask raster alone is
+  used.  The new mask raster must have the same dimensions and geotransform as
+  the output warped raster. https://github.com/natcap/pygeoprocessing/issues/366
 
 2.4.2 (2023-10-24)
 ------------------